Working on analyzer.
